VEOSS-1627 | Add check for existing previous registrations to prevent user deleting them. Optimise imports.

+++ app/controllers/previousRegistrations/DeletePreviousRegistrationController.scala
@@ -20,12 +20,13 @@ import controllers.actions._
 import forms.previousRegistrations.DeletePreviousRegistrationFormProvider
 import models.previousRegistrations.PreviousRegistrationDetailsWithOptionalVatNumber
 import models.requests.AuthenticatedDataRequest
-import models.{Index, Mode}
+import models.{AmendMode, Index, Mode}
 import pages.previousRegistrations.DeletePreviousRegistrationPage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents, Result}
-import queries.previousRegistration
+import queries.previousRegistration._
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
+import utils.CheckExistingRegistrations.{checkExistingRegistration, existingPreviousRegistration}
 import views.html.previousRegistrations.DeletePreviousRegistrationView
 
 import javax.inject.Inject
@@ -45,7 +46,14 @@ class DeletePreviousRegistrationController @Inject()(
     implicit request =>
       getPreviousRegistration(index) {
         details =>
-          Future.successful(Ok(view(form, mode, index, details.previousEuCountry.name)))
+
+          val existingPreviousRegistrations = checkExistingRegistration.previousRegistrations
+
+          if (mode == AmendMode && existingPreviousRegistration(details.previousEuCountry, existingPreviousRegistrations)) {
+            Future.successful(Redirect(controllers.routes.UnauthorisedController.onPageLoad())) //TODO new page?
+          } else {
+            Future.successful(Ok(view(form, mode, index, details.previousEuCountry.name)))
+          }
       }
   }
 
@@ -54,20 +62,27 @@ class DeletePreviousRegistrationController @Inject()(
       getPreviousRegistration(index) {
         details =>
 
-          form.bindFromRequest().fold(
-            formWithErrors =>
-              Future.successful(BadRequest(view(formWithErrors, mode, index, details.previousEuCountry.name))),
+          val existingPreviousRegistrations = checkExistingRegistration.previousRegistrations
 
-            value =>
-              if (value) {
-                for {
-                  updatedAnswers <- Future.fromTry(request.userAnswers.remove(previousRegistration.PreviousRegistrationQuery(index)))
-                  _              <- cc.sessionRepository.set(updatedAnswers)
-                } yield Redirect(DeletePreviousRegistrationPage(index).navigate(mode, updatedAnswers))
-              } else {
-                Future.successful(Redirect(DeletePreviousRegistrationPage(index).navigate(mode, request.userAnswers)))
-              }
-          )
+          if (mode == AmendMode && existingPreviousRegistration(details.previousEuCountry, existingPreviousRegistrations)) {
+            Future.successful(Redirect(controllers.routes.UnauthorisedController.onPageLoad()))//TODO new page?
+          } else {
+
+            form.bindFromRequest().fold(
+              formWithErrors =>
+                Future.successful(BadRequest(view(formWithErrors, mode, index, details.previousEuCountry.name))),
+
+              value =>
+                if (value) {
+                  for {
+                    updatedAnswers <- Future.fromTry(request.userAnswers.remove(PreviousRegistrationQuery(index)))
+                    _ <- cc.sessionRepository.set(updatedAnswers)
+                  } yield Redirect(DeletePreviousRegistrationPage(index).navigate(mode, updatedAnswers))
+                } else {
+                  Future.successful(Redirect(DeletePreviousRegistrationPage(index).navigate(mode, request.userAnswers)))
+                }
+            )
+          }
       }
   }
 
@@ -75,7 +90,7 @@ class DeletePreviousRegistrationController @Inject()(
   private def getPreviousRegistration(index: Index)
                              (block: PreviousRegistrationDetailsWithOptionalVatNumber => Future[Result])
                              (implicit request: AuthenticatedDataRequest[AnyContent]): Future[Result] =
-    request.userAnswers.get(previousRegistration.PreviousRegistrationWithOptionalVatNumberQuery(index)).map {
+    request.userAnswers.get(PreviousRegistrationWithOptionalVatNumberQuery(index)).map {
       details =>
         block(details)
     }.getOrElse(Future.successful(Redirect(controllers.routes.JourneyRecoveryController.onPageLoad())))
